loading state placeholders added
