+++ src/features/MoviePage/Content/Cast/index.js
@@ -4,32 +4,57 @@ import PersonTile from "../../../../common/Tiles/PersonTilesContainer/PersonTile
 import { TilesHeader } from "../../../../common/Tiles/TilesHeader/styled";
 import { TilesContainer } from "../../../../common/Tiles/PersonTilesContainer/styled";
 import noPersonImage from "../../images/no-person-image.png";
-import { selectMovieDetailsData } from "../../movieDetailsSlice";
+import {
+  selectMovieDetailsData,
+  selectMovieDetailsStatus,
+} from "../../movieDetailsSlice";
 
 const Cast = () => {
   const movieDetailsData = useSelector(selectMovieDetailsData);
+  const movieDetailsStatus = useSelector(selectMovieDetailsStatus);
+
   const imageBaseUrl = "https://image.tmdb.org/t/p/w185";
   const maxNumberOfTiles = 12;
 
   return (
     <>
-      <TilesHeader>Cast</TilesHeader>
-      <TilesContainer>
-        {movieDetailsData.credits.cast
-          .slice(0, maxNumberOfTiles)
-          .map((cast) => (
-            <PersonTile
-              key={nanoid()}
-              image={
-                cast.profile_path === null
-                  ? noPersonImage
-                  : imageBaseUrl + cast.profile_path
-              }
-              name={cast.name}
-              extraInfo={cast.character}
-            />
-          ))}
-      </TilesContainer>
+      {movieDetailsStatus === "loading" ? (
+        <>
+          <TilesHeader>Cast</TilesHeader>
+          <TilesContainer>
+            {[...Array(12)].map(() => (
+              <PersonTile
+                key={nanoid()}
+                image={noPersonImage}
+                name=""
+                extraInfo=""
+              />
+            ))}
+          </TilesContainer>
+        </>
+      ) : movieDetailsStatus === "done" ? (
+        <>
+          <TilesHeader>Cast</TilesHeader>
+          <TilesContainer>
+            {movieDetailsData.credits.cast
+              .slice(0, maxNumberOfTiles)
+              .map((cast) => (
+                <PersonTile
+                  key={nanoid()}
+                  image={
+                    cast.profile_path === null
+                      ? noPersonImage
+                      : imageBaseUrl + cast.profile_path
+                  }
+                  name={cast.name}
+                  extraInfo={cast.character}
+                />
+              ))}
+          </TilesContainer>
+        </>
+      ) : (
+        "ErrorPage"
+      )}
     </>
   );
 };
